Apply overrides for PangoFont using Module#prepend
