+++ lib/gir_ffi-pango.rb
@@ -4,6 +4,6 @@ require "gir_ffi"
 
 GirFFI.setup :Pango
 
-require "gir_ffi-pango/font"
+require "gir_ffi-pango/font_overrides"
 require "gir_ffi-pango/language"
 require "gir_ffi-pango/version"
